Staking rewards: only allow to claim if rewards > 0
